requirements.txt
streamlit
langchain
langchain-openai
langchain-community
pypdf
python-dotenv
faiss-cpu
